+++ example/lib/main.dart
@@ -1,4 +1,6 @@
+import 'package:adaptive_display/adaptive_display.dart';
 import 'package:flutter/material.dart';
+import 'package:widget_adaptive/widgets/adaptive_colorful.dart';
 import 'package:widget_adaptive/widgets/adaptive_stateful.dart';
 import 'package:widget_adaptive/widgets/adaptive_stateless.dart';
 import 'package:widget_adaptive/widgets/adaptive_widget.dart';
@@ -24,7 +26,7 @@ class MainApp extends StatelessWidget {
   }
 }
 
-class SelectorPage extends StatelessWidget {
+class SelectorPage extends AdaptiveWrapWidget {
   const SelectorPage({super.key});
 
   @override
@@ -53,10 +55,71 @@ class SelectorPage extends StatelessWidget {
               onPressed: () => _goToWidgetStateful(context),
               child: const Text('Widget Stateful'),
             ),
+            const SizedBox(height: 16),
+            ElevatedButton(
+              onPressed: () => _goToWidgetColorful(context),
+              child: const Text('Widget Colorful'),
+            ),
           ],
         ),
       ),
     );
+  }
+
+  @override
+  Widget md(BuildContext context, BoxConstraints constraints) {
+    return Scaffold(
+      body: Center(
+        child: Row(
+          children: [
+            Expanded(
+              child: ListView(
+                padding: const EdgeInsets.all(32),
+                children: [
+                  ElevatedButton(
+                    onPressed: () => _goToWidgetAdaptive(context),
+                    child: const Text('Widget Adaptive'),
+                  ),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                    onPressed: () => _goToWidgetWrap(context),
+                    child: const Text('Widget Adaptive Wrap'),
+                  ),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                    onPressed: () => _goToWidgetStateless(context),
+                    child: const Text('Widget Stateless'),
+                  ),
+                  const SizedBox(height: 16),
+                ],
+              ),
+            ),
+            Expanded(
+              child: ListView(
+                padding: const EdgeInsets.all(32),
+                children: [
+                  ElevatedButton(
+                    onPressed: () => _goToWidgetStateful(context),
+                    child: const Text('Widget Stateful'),
+                  ),
+                  const SizedBox(height: 16),
+                  ElevatedButton(
+                    onPressed: () => _goToWidgetColorful(context),
+                    child: const Text('Widget Colorful'),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  @override
+  Widget xs(BuildContext context, BoxConstraints constraints) {
+    // TODO: implement xs
+    throw UnimplementedError();
   }
 }
 
@@ -76,10 +139,15 @@ void _goToWidgetWrap(BuildContext context) {
   Navigator.pushNamed(context, '/widget_wrap');
 }
 
+void _goToWidgetColorful(BuildContext context) {
+  Navigator.pushNamed(context, '/widget_colorful');
+}
+
 final routes = {
   '/': (context) => const SelectorPage(),
   '/widget_adaptive': (context) => const WidgetAdaptive(),
   '/widget_wrap': (context) => const WidgetWrapAdaptive(),
   '/widget_stateless': (context) => const StateLessAdaptive(),
   '/widget_stateful': (context) => const StatefulAdaptive(),
+  '/widget_colorful': (context) => const ColorfulAdaptive(),
 };