+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Added getter for Adaptive settings in adaptive widgets.
+* Added a colorful example and improved the example's main
+
 ## 0.0.1
 
 * AdaptiveWrapState: Adaptive State for Stateful Widget.